droidmediacodec: refactoring buffer handling a bit.
Allocate a set of buffers to be used during the whole lifetime.
Remove _write() and split the process of passing buffers into dequeuing a buffer
and queuing a buffer. This allows us to easily discard data if we are not running.
The loop function is still started by us but it should be moved under the control
of the library user.