droidmediacamera: add a way to share the camera constants between droidmedia and its user.
If we need the user of droidmedia to not care about the Android version droidmedia is compiled
against then we have to provide the android camera constants in a way that does not need android
headers to be included by the library user.

This is the best way I can think of until now.