all: rework all our public API to get rid of booleans
If a function used to return a bool, it will now return an int. 0 is OK, anything else is not
If a function used to accept a bool, it will now accept an int. 1 is true and 0 is false.
Same goes for struct members.

Some functions used to query the status will return 1 as OK and 0 as false