[droidmedia] allow create droidmediabuffer from other formats
droid_media_buffer_create_from_raw_data assumes that the data is always
YV12. When other format is passed in, it can cause a crash.
This commit make sure that the function uses the correct number of lines
and stride for the format passed in.

This can happen if Gstreamer happens to negotiate raw NV21 from
droidcamsrc to droideglsink because the elements in the middle doesn't
support ENCODED format passthrough. While the proper fix needs to be
in Gstreamer (see [1]), this patch is still needed for correctness.

This commit also add a check to make sure only supported format is
passed in, and a logging message if calculated stride is too much
off.

[1] https://gitlab.freedesktop.org/gstreamer/gst-plugins-base/merge_requests/467