[camera] Add protection around DroidMediaCamera pointer in CameraListener
When destroying DroidMediaCamera, CameraListener held inside an ongoing
callback can outlive the destroying instance. To prevent it from
accessing DroidMediaCamera after being destroyed, one needs to remove
the reference to DroidMediaCamera inside the listener before deleting
it.

As this removal happens in a different thread then callbacks, a mutex is
required to prevent race condition. Also, just to be safe, remove the
listener from the Android Camera instance as well.