[droidmedia] Fix deadlock in droid_media_codec_stop()
The decoder will block in droid_media_codec_stop() if it is called before
EOS, but no futher frames are queued because the is no mechanism to
wake up the reader loop. After applying this commit, the read loop will be
notified of the state change in droid_media_codec_stop().