[droidmedia] Fix AsyncCodecSource to be asynchronous. JB#57676
When MediaCodec starts up, it spits a number of CB_INPUT_AVAILABLE
messages with indexes of available input buffers. The messages are
queued in message channel, so the output buffer will hang in queue
and will not be processed until all input buffers are filled with
data from MediaSource. It may take some time in case of a slow media
source, e.g. camera. This causes the output frame to be delayed by
12-14 frames.

This commit adds a SourceReader that stores available input buffer
indexes on order to unblock message queue and fetches the data from
the MediaSource in a separate thread.